refactor: decompose regression, data_augmentation, and debug_utils into focused submodules
Wave 3 of god-module decomposition targeting three 1000+ line files:

regression.py (1066 -> 324 lines facade):
- regression_types.py: enums (RegularizationType, SelectionMethod) and dataclasses
- regression_fitting.py: OLS, Ridge, LASSO, Elastic Net, feature selection (FittingMixin)
- regression_diagnostics.py: statistics, VIF, diagnostics, importance (DiagnosticsMixin)

data_augmentation.py (1066 -> 202 lines facade):
- augmentation_types.py: AugmentationMethod enum, config, result dataclasses
- augmentation_noise.py: Gaussian, uniform, colored, salt-pepper noise (NoiseMixin)
- augmentation_transforms.py: warping, scaling, rotation, SMOTE, mixup, frequency (TransformsMixin)

debug_utils.py (1123 -> 152 lines facade):
- debug_profiling.py: ProfileResult, TimingStats, profile/timer decorators
- debug_memory.py: MemoryStats, memory_tracker, memory_profile
- debug_tracing.py: ExecutionTracer, StackFrame, call stack inspection
- debug_helpers.py: logging, exceptions, diagnostics, watchdog, deprecation

All original modules serve as facades re-exporting public API for backward compatibility.
All existing tests pass (39/39 debug_utils, 102/103 data_processor - 1 pre-existing failure).
